backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
python-multipart
reportlab
